Remove typeahead (fixes #40, #111)
I initially started by trying to fix a regression to #80 which was introduced
in 2fbafa, but after having evaluated the code I discovered that there's no
easy way to both fix the regression and make the typeahead logic work again.
This is sadly a side-effect of no longer keeping `value` in state, since we
can no longer diff previous state and next state in the same fashion.

Since the current typeahead situation is very poor on mobile, not to mention
how broken it is in IE, I decided the best way forward is to temporarily remove
the typeahead logic.