Add isItemSelectable prop  [added] (#237)
Allow certain items to be determined unselectable (via mouse or keyboard), based on a user-defined boolean function prop of "isItemSelectable".  This allows for category headers or other external information to be included within the items list and evaluate to false in the determined function.  By default, all items return true.